Refactor member and team section handling by consolidating imports and removing obsolete configuration module
