+++ libs/commands/registry/__init__.py
@@ -1,7 +1,6 @@
 """
 メンバー/チーム操作モジュール
 
-- `libs.commands.registry.configuration`: 設定値み込み
 - `libs.commands.registry.member`: メンバー管理
 - `libs.commands.registry.team`: チーム管理
 """
