+++ libs/commands/registry/team.py
@@ -3,10 +3,112 @@ libs/registry/team.py
 """
 
 import logging
+from typing import TYPE_CHECKING, TypedDict
 
 import libs.global_value as g
+from libs.bootstrap.app_config import BaseSection
 from libs.data import initialization, lookup, modify
 from libs.utils import dbutil, formatter, textutil, validator
+
+if TYPE_CHECKING:
+    from libs.bootstrap.app_config import AppConfig
+
+
+class TeamDataDict(TypedDict):
+    """チーム情報格納辞書"""
+
+    id: int
+    """チームID"""
+    team: str
+    """チーム名"""
+    member: list[str]
+    """所属メンバーリスト"""
+
+
+class TeamSection(BaseSection):
+    """teamセクション処理"""
+
+    section: str
+    info: list[TeamDataDict]
+    """チーム情報"""
+    registration_limit: int
+    """登録チーム上限数"""
+    character_limit: int
+    """チーム名に使用できる文字数"""
+    member_limit: int
+    """チームに所属できるメンバー上限"""
+    friendly_fire: bool
+    """チームメイトが同卓しているゲームを集計対象に含めるか"""
+
+    def __init__(self):
+        self.section = "team"
+        self._reset()
+
+    def _reset(self):
+        self.info = []
+        self.registration_limit = int(255)
+        self.character_limit = int(16)
+        self.member_limit = int(16)
+        self.friendly_fire = bool(True)
+
+    def config_load(self, outer: "AppConfig"):
+        """設定値取り込み
+
+        Args:
+            outer (AppConfig): 設定クラスオブジェクト
+        """
+
+        self._parser = outer._parser
+        self._reset()
+        super().__init__(self)
+
+        # 呼び出しキーワード取り込み
+        self.commandword = [x.strip() for x in self._parser.get(self.section, "commandword", fallback="チーム一覧").split(",")]
+
+        logging.debug("%s: %s", self.section, self)
+
+    def member(self, team: str) -> list[str]:
+        """チーム所属メンバーをリストで返す
+
+        Args:
+            team (str): チーム名
+
+        Returns:
+            list[str]: 所属メンバーリスト
+        """
+
+        for x in self.info:
+            if x.get("team") == team:
+                return x.get("member")
+        return []
+
+    def which(self, name: str) -> str | None:
+        """指定メンバーの所属チームを返す
+
+        Args:
+            name (str): チェック対象のメンバー名
+
+        Returns:
+            Union[str, None]:
+            - str: 所属しているチーム名
+            - None: 未所属
+        """
+
+        for team in self.lists:
+            if name in self.member(team):
+                return team
+
+        return None
+
+    @property
+    def lists(self) -> list[str]:
+        """チーム名一覧をリストで返す
+
+        Returns:
+            list[str]: チーム名一覧
+        """
+
+        return [x.get("team") for x in self.info]
 
 
 def create(argument: list) -> str:
